refactor(rpc): move revm utils to module
